comment error처리, 리팩토링

+++ controller/comments.js
@@ -1,16 +1,16 @@
 const commentService = require("../services/comments");
 
+// 댓글 post 메소드
 const commentPost = async (req, res, next) => {
   try {
+    // params와 body에서값 받기 & body에서 오는 comment는 validate필요
     const { trackId } = req.params;
-    // comment validate만들기
     const { comment } = req.body;
+
+    // auth 미들웨어에서 오는 user정보 받기
     const { userId, nickname } = res.locals.user;
-    // validate 만들어야할거같은데
-    if (!comment || !trackId) {
-      res.sendStatus(400);
-      return;
-    }
+
+    // 댓글만드는 service 실행
     const result = await commentService.createComment({
       comment,
       trackId,
@@ -19,59 +19,50 @@ const commentPost = async (req, res, next) => {
     });
     return res.status(200).json({ comment: result });
   } catch (error) {
-    console.log(error);
+    error.status = 400;
     next(error);
   }
 };
 
+// 댓글 put 메소드
 const commentPut = async (req, res, next) => {
   try {
+    // params와 body에서값 받기 & body에서 오는 comment는 validate필요
     const { trackId, commentId } = req.params;
-    //comment validate만들기
     const { comment } = req.body;
+
+    // auth 미들웨어에서 오는 user정보 받기
     const { userId, nickname } = res.locals.user;
-    // middleware로 처리
-    if (!comment || !trackId || !commentId) {
-      res.sendStatus(400);
-      return;
-    }
-    // updateComment에 들어가는 객체도 class화 해서 만들자
-    const { createdAt, userId: commentUserId } = await commentService.findCommentByCommentId({
-      commentId,
-    });
-    if (userId !== commentUserId) {
-      res.sendStatus(400);
-      return;
-    }
+
+    // 댓글 수정하는 service 실행
     const result = await commentService.updateComment({
       commentId,
       comment,
+      userId,
       nickname,
-      createdAt,
       trackId,
     });
     return res.status(200).json({ comment: result });
   } catch (error) {
-    console.log(error);
+    error.status = 400;
     next(error);
   }
 };
 
+// 댓글 delete메소드
 const commentDelete = async (req, res, next) => {
   try {
+    // params에서값 받기
     const { trackId, commentId } = req.params;
-    const { userId } = res.locals.user;
-    const { userId: commentUserId } = await commentService.findCommentByCommentId({ commentId });
 
-    // 비교로직 utils에 만들자 throw error
-    if (userId !== commentUserId) {
-      res.sendStatus(400);
-      return;
-    }
-    await commentService.deleteComment({ trackId, commentId });
+    // auth 미들웨어에서 오는 user정보 받기
+    const { userId } = res.locals.user;
+
+    // 댓글 삭제하는 service 실행
+    await commentService.deleteComment({ userId, trackId, commentId });
     return res.sendStatus(200);
   } catch (error) {
-    console.log(error);
+    error.status = 400;
     next(error);
   }
 };
